sem2/Lab03/requirements.txt
PyQt5
msgpack
pycryptodome
pyOpenSSL